cm2html: Process "pandoc"-style header into Dublin Core <META> elements and the <TITLE>.
If the first (up to) three lines of the input field have a `%` in the first column:

    % The Title
    % A. Uthor
    % 2015-10-18

then the document title is set from the first line, the author and date from the following lines;
and these items are put into the `<HEAD>` in `<META>` elements according to the Dublin Core.

The title is also the document `<TITLE>`, unless overwritten by the `-t` option.
If no title is given in either way, the string "Untitled Document" is used.