More hackery in `scanners.re`: Allow `<gi 0>` and `<gi att=val/`
The abominable HTML5 "syntax" considers an attribute spec without
the "=" value part to be an *attribute name*, not an *attribute value*
(a NMTOKEN). So `<gi 0>` would see `0` as an (invalid?) *attribute name*.

In "unquoted" attribute values a `/` is actually treated as part
of the "value" (which is NOT a NMTOKEN ...), so `<gi att=foo/>` is
a start tag with "foo/" as an attribute value, NOT an empty-element
tag. Similarly, `<gi att=foo/` is NOT a NET-enabling start tag, but
extends happily beyond the `/` ...

WTF were they thinking?